+++ src/main/java/com/autogeneral/codetest/dao/ToDoRepository.java
@@ -3,5 +3,5 @@ package com.autogeneral.codetest.dao;
 import com.autogeneral.codetest.model.ToDo;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface ToDoRepository extends PagingAndSortingRepository<ToDo, Integer> {
+public interface ToDoRepository extends PagingAndSortingRepository<ToDo, Long> {
 }
